# workspaces/x2a/plugins/x2a-backend/src/schema/openapi.yaml
openapi: 3.0.4
info:
  title: x2a
  description: The [X2Ansible](https://github.com/redhat-developer/x2ansible) API.
  version: 0.0.1
servers:
  - url: /
paths:
...
  /projects/{projectId}/collectArtifacts:
    post:
      summary: Collects artifacts from a completed X2Ansible job
      description: |
        Callback endpoint for X2Ansible jobs to submit execution artifacts and results.
        This endpoint is called by the X2Ansible job runner when a migration phase completes.
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          required: true
          description: UUID of the project
        - in: query
          name: moduleId
          schema:
            type: string
          required: false
          description: |
            UUID of the module.
            - Required for analyze, migrate, and publish phases
            - Should be omitted for init phase
        - in: query
          name: phase
          schema:
            $ref: '#/components/schemas/MigrationPhase'
          required: true
          description: Migration phase that completed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - Success
                    - Error
                  description: Execution status of the job
                errorDetails:
                  type: string
                  description: Error details if status is Error
                jobId:
                  type: string
                  description: UUID of the completed job
                artifacts:
                  type: array
                  description: List of artifacts produced by the job
                  items:
                    $ref: '#/components/schemas/Artifact'
                telemetry:
                  $ref: '#/components/schemas/Telemetry'
              required:
                - status
                - jobId
                - artifacts
      responses:
        '200':
          description: Artifacts collected successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success confirmation message
                required:
                  - message
        '404':
          description: Project or job not found
        '400':
          description: Invalid request data
components:
  schemas:
...
    ArtifactType:
      type: string
      enum:
        - migration_plan
        - module_migration_plan
        - migrated_sources
    Artifact:
      type: object
      properties:
        id:
          type: string
          description: UUID for the artifact
        type:
          $ref: '#/components/schemas/ArtifactType'
          description: Type of the artifact
        value:
          type: string
          description: Value of the artifact
      required:
        - id
        - type
        - value
...
    MigrationPhase:
      type: string
      enum:
        - init
        - analyze
        - migrate
        - publish
      description: All migration phases
...
    Telemetry:
      type: object
      description: Execution telemetry and metrics for a migration phase
      properties:
        summary:
          type: string
          description: Execution summary from x2aconvertor
        phase:
          type: string
          description: Phase name (init, analyze, migrate, publish)
        startedAt:
          type: string
          format: date-time
          description: When the phase started (ISO 8601 format)
        endedAt:
          type: string
          format: date-time
          description: When the phase completed (ISO 8601 format)
        agents:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AgentMetrics'
          description: Per-agent execution metrics (agent name to metrics mapping)
      required:
        - summary
        - phase
        - startedAt
    AgentMetrics:
      type: object
      description: Telemetry data for a single agent execution within a phase
      properties:
        name:
          type: string
          description: Agent identifier (e.g., "PlanningAgent", "WriteAgent")
        startedAt:
          type: string
          format: date-time
          description: When agent execution started (ISO 8601 format)
        endedAt:
          type: string
          format: date-time
          description: When agent execution completed (ISO 8601 format)
        durationSeconds:
          type: number
          format: double
          description: Execution duration in seconds
        metrics:
          type: object
          additionalProperties: true
          description: Custom key-value metrics recorded by the agent
        toolCalls:
          type: object
          additionalProperties:
            type: integer
          description: Tool call counts (tool name to count mapping)
      required:
        - name
        - durationSeconds